OCPBUGS-34077: Always Disable Default Rolebindings Controller
With OCPBUILD-7, default namespace rolebindings for the Image Registry,
DeploymentConfig, and Builds are handled by separate controllers. The
initial refactor in OCPBUILD-8 kept the default rolebindings controller
in place to allow upgrade from 4.15 -> 4.16. However, the corresponding
operator change in OCPBUILD-9 forgot to disable the default
rolebindings controller. This leads to the following side effects:

- On clusters with no capabilities, the default role bindings are
  created and reconciled.
- On a standard installation, there is potential for race conditions
  and server-side conflicts as two or more controllers fight to
  reconcile a resource. At scale, this can lead to excessive load on
  the apiserver.

This change ensures that the default rolebindings controller is always
disabled. As a follow-up, the code for the default rolebindings
controller can be safely removed in OCP 4.17.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>